Add reduction op, still need tests
